+++ prt/helper.cpp
@@ -1,0 +1,115 @@
+#include <assert.h>
+#ifdef WINDOWS
+#include <windows.h>
+#endif
+#include <sys/stat.h>
+#include <string>
+#include <stdio.h>
+
+// Error handler: print message if any, and exit
+void quit(const char* message=0) {
+  throw message;
+}
+
+// strings are equal ignoring case?
+int equals(const char* a, const char* b) {
+  assert(a && b);
+  while (*a && *b) {
+    int c1=*a;
+    if (c1>='A'&&c1<='Z') c1+='a'-'A';
+    int c2=*b;
+    if (c2>='A'&&c2<='Z') c2+='a'-'A';
+    if (c1!=c2) return 0;
+    ++a;
+    ++b;
+  }
+  return *a==*b;
+}
+
+bool makedir(const char* dir) {
+  struct stat status;
+  bool success = stat(dir, &status)==0;
+  if(success && (status.st_mode & S_IFDIR)!=0) return -1; //-1: directory already exists, no need to create
+#ifdef WINDOWS
+  return CreateDirectory(dir, 0)==TRUE;
+#else
+#ifdef UNIX
+  return mkdir(dir, 0777)==0;
+#else
+  return false;
+#endif
+#endif
+}
+
+void makedirectories(const char* filename) {
+  std::string path(filename);
+  int start = 0;
+  if(path[1]==':')start=2; //skip drive letter (c:)
+  if(path[start] == '/' || path[start] == '\\')start++; //skip leading slashes (root dir)
+  for (int i = start; path[i]; ++i) {
+    if (path[i] == '/' || path[i] == '\\') {
+      char savechar = path[i];
+      path[i] = 0;
+      const char* dirname = path.c_str();
+      int created = makedir(dirname);
+      if (created==0) {
+        printf("Unable to create directory %s\n", dirname);
+        quit();
+      }
+      if(created==1)
+       // printf("Created directory %s\n", dirname);
+      path[i] = savechar;
+    }
+  }
+}
+
+/////////////////////////// File /////////////////////////////
+// The main purpose of these classes is to keep temporary files in 
+// RAM as mush as possible. The default behaviour is to simply pass 
+// function calls to the operating system - except in case of temporary 
+// files.
+
+// Helper function: create a temporary file
+//
+// On Windows when using tmpfile() the temporary file may be created 
+// in the root directory causing access denied error when User Account Control (UAC) is on.
+// To avoid this issue with tmpfile() we simply use fopen() instead.
+// We create the temporary file in the directory where the executable is launched from. 
+// Luckily the MS C runtime library provides two (MS specific) fopen() flags: "T"emporary and "D"elete.
+
+
+FILE* tmpfile2(void){
+    FILE *f;
+#if defined(WINDOWS)    
+    int i;
+    char temppath[MAX_PATH]; 
+    char filename[MAX_PATH];
+    
+    //i=GetTempPath(MAX_PATH,temppath);          //store temp file in system temp path
+    i=GetModuleFileName(NULL,temppath,MAX_PATH); //store temp file in program folder
+    if ((i==0) || (i>MAX_PATH)) return NULL;
+    char *p=strrchr(temppath, '\\');
+    if (p==0) return NULL;
+    p++;*p=0;
+    if (GetTempFileName(temppath,"tmp",0,filename)==0) return NULL;
+    f=fopen(filename,"w+bTD");
+    if (f==NULL) unlink(filename);
+    return f;
+#else
+    f=tmpfile();  // temporary file
+    if (!f) return NULL;
+    return f;
+#endif
+}
+/*
+FILE* maketmpfile(void) {
+#if defined(WINDOWS)  
+  char szTempFileName[MAX_PATH];
+  UINT uRetVal = GetTempFileName(TEXT("."), TEXT("tmp"), 0, szTempFileName);
+  if(uRetVal==0)return 0;
+  return fopen(szTempFileName, "w+bTD");
+#else
+  return tmpfile();
+#endif
+}*/
+
