+++ prt/log.hpp
@@ -1,0 +1,41 @@
+#pragma once
+#include "array.hpp"
+
+int ilog(U16 x);
+int llog(U32 x);
+ 
+inline U32 BitCount(U32 v) {
+  v -= ((v >> 1) & 0x55555555);
+  v = ((v >> 2) & 0x33333333) + (v & 0x33333333);
+  v = ((v >> 4) + v) & 0x0f0f0f0f;
+  v = ((v >> 8) + v) & 0x00ff00ff;
+  v = ((v >> 16) + v) & 0x0000ffff;
+  return v;
+}
+
+// ilog2
+// returns floor(log2(x)), e.g. 30->4  31->4  32->5,  33->5
+#ifdef _MSC_VER
+#include <intrin.h>
+inline U32 ilog2(U32 x) {
+  DWORD tmp=0;
+  if(x!=0)_BitScanReverse(&tmp,x);
+  return tmp;
+}
+#elif __GNUC__
+inline U32 ilog2(U32 x) {
+  if(x!=0)x=31-__builtin_clz(x);
+  return x;
+}
+#else
+inline U32 ilog2(U32 x) {
+  //copy the leading "1" bit to its left (0x03000000 -> 0x03ffffff)
+  x |= (x >> 1);
+  x |= (x >> 2);
+  x |= (x >> 4);
+  x |= (x >> 8);
+  x |= (x >>16);
+  //how many trailing bits do we have (except the first)? 
+  return BitCount(x >> 1);
+}
+#endif
