+++ prt/logistic.cpp
@@ -1,0 +1,61 @@
+#include "logistic.hpp"
+
+///////////////////////////// Squash //////////////////////////////
+// return p = 1/(1 + exp(-d)), d scaled by 8 bits, p scaled by 12 bits
+class Squash {
+  Array<U16> t;
+public:
+  Squash();
+  int operator()(int p) const {
+   if (p<-2047) return  0; 
+   if (p>2047) return  4095;
+   return t[p+2048];
+  }
+} sqh;
+
+Squash::Squash(): t(4096) {
+int ts[33]={1,2,3,6,10,16,27,45,73,120,194,310,488,747,1101,
+    1546,2047,2549,2994,3348,3607,3785,3901,3975,4022,
+    4050,4068,4079,4085,4089,4092,4093,4094};
+  int w,d;
+  for (int i=-2047; i<=2047; ++i){
+    w=i&127;
+  d=(i>>7)+16;
+  t[i+2048]=(ts[d]*(128-w)+ts[(d+1)]*w+64) >> 7;
+    }
+}
+
+int squash(int p)  {
+   return sqh(p);
+  
+} 
+//////////////////////////// Stretch ///////////////////////////////
+
+// Inverse of squash. d = ln(p/(1-p)), d scaled by 8 bits, p by 12 bits.
+// d has range -2047 to 2047 representing -8 to 8.  p has range 0 to 4095.
+
+class Stretch {
+  Array<short> t;
+public:
+  Stretch();
+  int operator()(int p) const {
+    assert(p>=0 && p<4096);
+    return t[p];
+  }
+} str;
+
+Stretch::Stretch(): t(4096) {
+  int pi=0;
+  for (int x=-2047; x<=2047; ++x) {  // invert squash()
+    int i=squash(x);
+    for (int j=pi; j<=i; ++j)
+      t[j]=x;
+    pi=i+1;
+  }
+  t[4095]=2047;
+}
+
+int stretch(int p) {
+ return str(p);
+}
+
