+++ prt/logistic.hpp
@@ -1,0 +1,5 @@
+#pragma once
+#include "types.hpp"
+#include "array.hpp"
+int squash(int p);
+int stretch(int p);
