+++ prt/mixer.hpp
@@ -1,0 +1,252 @@
+#pragma once
+#include "types.hpp"
+#include "array.hpp"
+#include "blockdata.hpp"
+#include "log.hpp"
+#include "logistic.hpp"
+
+//////////////////////////// Mixer /////////////////////////////
+
+// Mixer m(N, M, S=1, w=0) combines models using M neural networks with
+//   N inputs each, of which up to S may be selected.  If S > 1 then
+//   the outputs of these neural networks are combined using another
+//   neural network (with parameters S, 1, 1).  If S = 1 then the
+//   output is direct.  The weights are initially w (+-32K).
+//   It is used as follows:
+// m.update() trains the network where the expected output is the
+//   last bit (in the global variable y).
+// m.add(stretch(p)) inputs prediction from one of N models.  The
+//   prediction should be positive to predict a 1 bit, negative for 0,
+//   nominally +-256 to +-2K.  The maximum allowed value is +-32K but
+//   using such large values may cause overflow if N is large.
+// m.set(cxt, range) selects cxt as one of 'range' neural networks to
+//   use.  0 <= cxt < range.  Should be called up to S times such
+//   that the total of the ranges is <= M.
+// m.p() returns the output prediction that the next bit is 1 as a
+//   12 bit number (0 to 4095).
+
+#if !defined(__GNUC__)
+
+#if (2 == _M_IX86_FP) // 2 if /arch:SSE2 was used.
+# define __SSE2__
+#elif (1 == _M_IX86_FP) // 1 if /arch:SSE was used.
+# define __SSE__
+#endif
+
+#endif /* __GNUC__ */
+
+#if defined(__AVX2__)
+#include <immintrin.h>
+#define OPTIMIZE "AVX2-"
+#elif defined(__SSE4_1__)   
+#include<smmintrin.h>
+#elif   defined(__SSSE3__)
+#include<tmmintrin.h>
+#elif defined(__SSE2__) 
+#include <emmintrin.h>
+#define OPTIMIZE "SSE2-"
+
+#elif defined(__SSE__)
+#include <xmmintrin.h>
+#define OPTIMIZE "SSE-"
+#endif
+/**
+ * Vector product a*b of n signed words, returning signed integer scaled down by 8 bits.
+ * n is rounded up to a multiple of 8.
+ */
+//static int dot_product (const short* const t, const short* const w, int n);
+
+/**
+ * Train n neural network weights w[n] on inputs t[n] and err.
+ * w[i] += ((t[i]*2*err)+(1<<16))>>17 bounded to +- 32K.
+ * n is rounded up to a multiple of 8.
+ */
+//static void train (const short* const t, short* const w, int n, const int e);
+struct ErrorInfo {
+  U32 Data[2], Sum, Mask, Collected;
+};
+
+
+
+#if defined(__MMX__)
+typedef __m128i XMM;
+#endif
+#if defined(__AVX2__)
+typedef __m256i YMM;
+#endif
+#define DEFAULT_LEARNING_RATE 7
+class Mixer {
+private: 
+  const int N, M, S;   // max inputs, max contexts, max context sets
+  Array<short, 32> tx; // N inputs from add()  
+  Array<short, 32> wx; // N*M weights
+  Array<U32> cxt;  // S contexts
+  Array<int> mrate;  // S contexts
+  Array<int> merr; 
+  int ncxt;        // number of contexts (0 to S)
+  int base;        // offset of next context
+public:
+  int nx;          // Number of inputs in tx, 0 to N  
+private:
+  Mixer* mp;       // points to a Mixer to combine results
+  Array<int> pr;   // last result (scaled 12 bits)
+  Array<ErrorInfo> info; 
+  Array<int> rates; // learning rates
+  int lrate,lshift; 
+public:  
+  BlockData& x;
+  Mixer(int n, int m,BlockData& bd, int s=1, int w=0,int g=7,int h=1);
+  
+#if defined(__AVX2__)
+ int dot_product (const short* const t, const short* const w, int n) {
+  assert(n == ((n + 15) & -16));
+  __m256i sum = _mm256_setzero_si256 ();
+  while ((n -= 16) >= 0) { // Each loop sums 16 products
+    __m256i tmp = _mm256_madd_epi16 (*(__m256i *) &t[n], *(__m256i *) &w[n]); // t[n] * w[n] + t[n+1] * w[n+1]
+    tmp = _mm256_srai_epi32 (tmp, 8); //                                        (t[n] * w[n] + t[n+1] * w[n+1]) >> 8
+    sum = _mm256_add_epi32 (sum, tmp); //                                sum += (t[n] * w[n] + t[n+1] * w[n+1]) >> 8
+  } 
+   sum =_mm256_hadd_epi32(sum,_mm256_setzero_si256 ());       //add [1]=[1]+[2], [2]=[3]+[4], [3]=0, [4]=0, [5]=[5]+[6], [6]=[7]+[8], [7]=0, [8]=0
+   sum =_mm256_hadd_epi32(sum,_mm256_setzero_si256 ());       //add [1]=[1]+[2], [2]=0,       [3]=0, [4]=0, [5]=[5]+[6], [6]=0,       [7]=0, [8]=0
+   __m128i lo = _mm256_extractf128_si256(sum, 0);
+   __m128i hi = _mm256_extractf128_si256(sum, 1);
+   __m128i newsum = _mm_add_epi32(lo, hi);                    //sum last two
+   return _mm_cvtsi128_si32(newsum);
+}
+
+ void train (const short* const t, short* const w, int n, const int e) {
+  assert(n == ((n + 15) & -16));
+  if (e) {
+    const __m256i one = _mm256_set1_epi16 (1);
+    const __m256i err = _mm256_set1_epi16 (short(e));
+    while ((n -= 16) >= 0) { // Each iteration adjusts 16 weights
+      __m256i tmp = _mm256_adds_epi16 (*(__m256i *) &t[n], *(__m256i *) &t[n]); // t[n] * 2
+      tmp = _mm256_mulhi_epi16 (tmp, err); //                                     (t[n] * 2 * err) >> 16
+      tmp = _mm256_adds_epi16 (tmp, one); //                                     ((t[n] * 2 * err) >> 16) + 1
+      tmp = _mm256_srai_epi16 (tmp, 1); //                                      (((t[n] * 2 * err) >> 16) + 1) >> 1
+      tmp = _mm256_adds_epi16 (tmp, *(__m256i *) &w[n]); //                    ((((t[n] * 2 * err) >> 16) + 1) >> 1) + w[n]
+      *(__m256i *) &w[n] = tmp; //                                          save the new eight weights, bounded to +- 32K
+    }
+  }
+}
+
+#elif defined(__SSE2__) || defined(__SSSE3__)
+ int dot_product (const short* const t, const short* const w, int n) {
+  assert(n == ((n + 15) & -16));
+  XMM sum = _mm_setzero_si128 ();
+  while ((n -= 8) >= 0) { // Each loop sums eight products
+    XMM tmp = _mm_madd_epi16 (*(XMM *) &t[n], *(XMM *) &w[n]); // t[n] * w[n] + t[n+1] * w[n+1]
+    tmp = _mm_srai_epi32 (tmp, 8); //                                        (t[n] * w[n] + t[n+1] * w[n+1]) >> 8
+    sum = _mm_add_epi32 (sum, tmp); //                                sum += (t[n] * w[n] + t[n+1] * w[n+1]) >> 8
+  }
+  sum = _mm_add_epi32(sum, _mm_srli_si128 (sum, 8));
+  sum = _mm_add_epi32(sum, _mm_srli_si128 (sum, 4));
+  return _mm_cvtsi128_si32 (sum); //                     ...  and scale back to integer
+}
+
+ void train (const short* const t, short* const w, int n, const int e) {
+  assert(n == ((n + 15) & -16));
+  if (e) {
+    const XMM one = _mm_set1_epi16 (1);
+    const XMM err = _mm_set1_epi16 (short(e));
+    while ((n -= 8) >= 0) { // Each iteration adjusts eight weights
+      XMM tmp = _mm_adds_epi16 (*(XMM *) &t[n], *(XMM *) &t[n]); // t[n] * 2
+      tmp = _mm_mulhi_epi16 (tmp, err); //                                     (t[n] * 2 * err) >> 16
+      tmp = _mm_adds_epi16 (tmp, one); //                                     ((t[n] * 2 * err) >> 16) + 1
+      tmp = _mm_srai_epi16 (tmp, 1); //                                      (((t[n] * 2 * err) >> 16) + 1) >> 1
+      tmp = _mm_adds_epi16 (tmp, *(XMM *) &w[n]); //                    ((((t[n] * 2 * err) >> 16) + 1) >> 1) + w[n]
+      *(XMM *) &w[n] = tmp; //                                          save the new eight weights, bounded to +- 32K
+    }
+  }
+}
+
+#elif defined(__SSE__)
+ int dot_product (const short* const t, const short* const w, int n) {
+  assert(n == ((n + 15) & -16));
+  __m64 sum = _mm_setzero_si64 ();
+  while ((n -= 8) >= 0) { // Each loop sums eight products
+    __m64 tmp = _mm_madd_pi16 (*(__m64 *) &t[n], *(__m64 *) &w[n]); //   t[n] * w[n] + t[n+1] * w[n+1]
+    tmp = _mm_srai_pi32 (tmp, 8); //                                    (t[n] * w[n] + t[n+1] * w[n+1]) >> 8
+    sum = _mm_add_pi32 (sum, tmp); //                            sum += (t[n] * w[n] + t[n+1] * w[n+1]) >> 8
+
+    tmp = _mm_madd_pi16 (*(__m64 *) &t[n + 4], *(__m64 *) &w[n + 4]); // t[n+4] * w[n+4] + t[n+5] * w[n+5]
+    tmp = _mm_srai_pi32 (tmp, 8); //                                    (t[n+4] * w[n+4] + t[n+5] * w[n+5]) >> 8
+    sum = _mm_add_pi32 (sum, tmp); //                            sum += (t[n+4] * w[n+4] + t[n+5] * w[n+5]) >> 8
+  }
+  sum = _mm_add_pi32 (sum, _mm_srli_si64 (sum, 32)); // Add eight sums together ...
+  const int retval = _mm_cvtsi64_si32 (sum); //                     ...  and scale back to integer
+  _mm_empty(); // Empty the multimedia state
+  return retval;
+}
+
+ void train (const short* const t, short* const w, int n, const int e) {
+  assert(n == ((n + 15) & -16));
+  if (e) {
+    const __m64 one = _mm_set1_pi16 (1);
+    const __m64 err = _mm_set1_pi16 (short(e));
+    while ((n -= 8) >= 0) { // Each iteration adjusts eight weights
+      __m64 tmp = _mm_adds_pi16 (*(__m64 *) &t[n], *(__m64 *) &t[n]); //   t[n] * 2
+      tmp = _mm_mulhi_pi16 (tmp, err); //                                 (t[n] * 2 * err) >> 16
+      tmp = _mm_adds_pi16 (tmp, one); //                                 ((t[n] * 2 * err) >> 16) + 1
+      tmp = _mm_srai_pi16 (tmp, 1); //                                  (((t[n] * 2 * err) >> 16) + 1) >> 1
+      tmp = _mm_adds_pi16 (tmp, *(__m64 *) &w[n]); //                  ((((t[n] * 2 * err) >> 16) + 1) >> 1) + w[n]
+      *(__m64 *) &w[n] = tmp; //                                       save the new four weights, bounded to +- 32K
+
+      tmp = _mm_adds_pi16 (*(__m64 *) &t[n + 4], *(__m64 *) &t[n + 4]); // t[n+4] * 2
+      tmp = _mm_mulhi_pi16 (tmp, err); //                                 (t[n+4] * 2 * err) >> 16
+      tmp = _mm_adds_pi16 (tmp, one); //                                 ((t[n+4] * 2 * err) >> 16) + 1
+      tmp = _mm_srai_pi16 (tmp, 1); //                                  (((t[n+4] * 2 * err) >> 16) + 1) >> 1
+      tmp = _mm_adds_pi16 (tmp, *(__m64 *) &w[n + 4]); //              ((((t[n+4] * 2 * err) >> 16) + 1) >> 1) + w[n]
+      *(__m64 *) &w[n + 4] = tmp; //                                   save the new four weights, bounded to +- 32K
+    }
+    _mm_empty(); // Empty the multimedia state
+  }
+}
+#else
+
+// dot_product returns dot product t*w of n elements.  n is rounded
+// up to a multiple of 8.  Result is scaled down by 8 bits.
+int dot_product(short *t, short *w, int n) {
+  int sum=0;
+  n=(n+15)&-16;
+  for (int i=0; i<n; i+=2)
+    sum+=(t[i]*w[i]+t[i+1]*w[i+1]) >> 8;
+  return sum;
+}
+
+// Train neural network weights w[n] given inputs t[n] and err.
+// w[i] += t[i]*err, i=0..n-1.  t, w, err are signed 16 bits (+- 32K).
+// err is scaled 16 bits (representing +- 1/2).  w[i] is clamped to +- 32K
+// and rounded.  n is rounded up to a multiple of 8.
+
+void train(short *t, short *w, int n, int err) {
+  n=(n+15)&-16;
+  for (int i=0; i<n; ++i) {
+    int wt=w[i]+(((t[i]*err*2>>16)+1)>>1);
+    if (wt<-32768) wt=-32768;
+    if (wt>32767) wt=32767;
+    w[i]=wt;
+  }
+}
+#endif 
+
+  // Adjust weights to minimize coding cost of last prediction
+  void update();
+  void update1();
+  void reset() {
+    nx=base=ncxt=0;
+  }
+  void update2();
+  // Input x (call up to N times)
+  void add(int x);
+  void sp(int x);
+  // Set a context (call S times, sum of ranges <= M)
+  void set(int cx, int range,int mr=14,int me=7*2);
+  void setl(int l, int r);
+
+  // predict next bit
+  int p(const int shift0=0, const int shift1=0);
+  ~Mixer();
+};
+
+
