+++ prt/rnd.hpp
@@ -1,0 +1,17 @@
+//////////////////////////// rnd ///////////////////////////////
+
+// 32-bit pseudo random number generator
+class Random{
+  Array<U32> table;
+  int i;
+public:
+  Random(): table(64) {
+    table[0]=123456789;
+    table[1]=987654321;
+    for (int j=0; j<62; j++) table[j+2]=table[j+1]*11+table[j]*23/16;
+    i=0;
+  }
+  U32 operator()() {
+    return ++i, table[i&63]=table[(i-24)&63]^table[(i-55)&63];
+  }
+} ;
