+++ prt/types.hpp
@@ -1,0 +1,16 @@
+#pragma once
+#include <stdint.h>
+#ifdef _MSC_VER
+
+typedef __int32 int32_t;
+typedef unsigned __int32 uint32_t;
+typedef __int64 int64_t;
+typedef unsigned __int64 uint64_t;
+
+#endif
+// 8, 16, 32 bit unsigned types (adjust as appropriate)
+typedef unsigned char  U8;
+typedef unsigned short U16;
+typedef unsigned int   U32;
+typedef uint64_t   U64;
+typedef signed char int8_t;
